+++ packages/black-forest-labs/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ai-sdk/black-forest-labs
 
+## 2.0.40
+
+### Patch Changes
+
+- Updated dependencies [9942196]
+  - @ai-sdk/provider@4.0.13
+  - @ai-sdk/provider-utils@5.0.39
+
+## 2.0.39
+
+### Patch Changes
+
+- Updated dependencies [912fb01]
+  - @ai-sdk/provider@4.0.12
+  - @ai-sdk/provider-utils@5.0.38
+
 ## 2.0.38
 
 ### Patch Changes