+++ packages/fal/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ai-sdk/fal
 
+## 3.0.40
+
+### Patch Changes
+
+- Updated dependencies [9942196]
+  - @ai-sdk/provider@4.0.13
+  - @ai-sdk/provider-utils@5.0.39
+
+## 3.0.39
+
+### Patch Changes
+
+- Updated dependencies [912fb01]
+  - @ai-sdk/provider@4.0.12
+  - @ai-sdk/provider-utils@5.0.38
+
 ## 3.0.38
 
 ### Patch Changes