+++ packages/otel/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ai-sdk/otel
 
+## 1.0.97
+
+### Patch Changes
+
+- Updated dependencies [ef3bac4]
+- Updated dependencies [9942196]
+  - ai@7.0.97
+  - @ai-sdk/provider@4.0.13
+
+## 1.0.96
+
+### Patch Changes
+
+- Updated dependencies [912fb01]
+- Updated dependencies [c595e6e]
+  - @ai-sdk/provider@4.0.12
+  - ai@7.0.96
+
 ## 1.0.95
 
 ### Patch Changes